requirements.txt
numpy>=1.19.5
torch>=0.3.0
torchvision>=0.2.0
Pillow>=8.0.0
pyyaml>=5.1
tqdm>=4.62.0
matplotlib>=3.3.0
scikit-image>=0.17.2
opencv-python>=4.5.0
tensorboard>=2.4.0 